requirements.txt
asgiref==3.8.1
boto3==1.37.8
botocore==1.37.8
Django==5.1.7
django-redis==5.4.0
django-storages==1.14.5
djangorestframework==3.15.2
gunicorn==23.0.0
jmespath==1.0.1
packaging==24.2
pillow==11.1.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyotp==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.2.1
s3transfer==0.11.4
six==1.17.0
sqlparse==0.5.3
urllib3==2.3.0
